requirements.txt
numpy
pillow
opencv-python
scikit-image
matplotlib
tabulate
pycocotools
torchvision
